backend-arbitragem/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
supabase==2.3.6
openai==1.13.3
beautifulsoup4==4.12.3
playwright==1.42.0
pydantic==2.6.3
python-dotenv==1.0.1
